+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feat(view.Auth): Logout users after a defined amount of time
 - feat(template.\*): Improve error style display
+- feat(forms.\*): Add toggle button on forms
 
 ### Changed